requirements.txt
streamlit
PyPDF2
pydantic
requests
python-dotenv
google-generativeai